requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
requests==2.32.3
openai==1.51.2
python-dotenv==1.0.1
pandas==2.2.3
tabulate==0.9.0
